move toMatchImageSnapshot closer to where it's used
